rollback meaningless commits

+++ backbones/resnet_blocks.py
@@ -199,7 +199,7 @@ class BlockType2(Keras3_Model_Wrapper):
         x = self.conv3_conv(x)
         x = self.conv3_bn(x, training=training)
 
-        x = tf.add(tf.cast(shortcut, x.dtype), x)
+        x = tf.add(shortcut, x)
         x = tf.nn.relu(x)
 
         return x
